### What I did

before:

```
In [1]: %time import ape_blast
CPU times: user 1.13 s, sys: 173 ms, total: 1.3 s
Wall time: 1.53 s
```

after:

```
In [1]: %time import ape_blast
CPU times: user 8.85 ms, sys: 4.13 ms, total: 13 ms
Wall time: 14.2 ms
```
### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)